Update OrdinalNumbers.js
Your task is to write the ordinal(number, brief) function. number will be an integer. You need to find the ordinal suffix of said number.
brief is an optional parameter and defaults to false. When using brief notation, nd and rd use d instead. All others are the same.
ordinal(number, brief) should return a string containing those two characters (or one character) that would be tagged onto the end of the number.
The last two digits determine the ordinal suffix.