Point in Polygon
In this kata, you're going write a function called pointInPoly to test if a point is inside a polygon.
Points will be represented as [x,y] arrays.
The polygon will be an array of points which are the polygon's vertices. The last point in the array connects back to the first point.