Two Sets of Equal Sum
Input
A positive integer n <= 1,000,000.

Output
If it's not possible, return [ ] (Javascript and Python) or ( ) (Python) or [[],[] ] (Java, C#, Kotlin) or None (Scala). If it's possible, return two disjoint sets. Each integer from 1 to n must be in one of them. The integers in the first set must sum up to the same value as the integers in the second set. The sets can be returned in a tuple, list, or array, depending on language.