Steps in k-primes
We will write a function kprimes_step(k, step, start, nd) with parameters:
k (integer > 0) which indicates the type of k-primes we are looking for,
step (integer > 0) which indicates the step we want to find between two k-primes,
start (integer >= 0) which gives the start of the search (start inclusive),
nd (integer >= start) which gives the end of the search (nd inclusive)
In the example above kprimes_step(2, 2, 0, 50) will return  [[4, 6], [33, 35]] which are the pairs of 2-primes between 2 and 50 with a 2 steps.
So this function should return an array of all the pairs (or tuples) of k-prime numbers spaced with a step of step between the limits start, nd. This array can be empty.