Simple Fun #110: Array Operations
You are given an array of integers a and a non-negative number of operations k, applied to the array. Each operation consists of two parts:

find the maximum element value of the array;
replace each element a[i] with (maximum element value - a[i]).```
How will the array look like after `k` such operations?